Add files via upload
In this file, in order to achieve the research's results, I used network analysis. Due to the problem that I had only information about artists, and not about genres, I deciced to group the artists for the respective genre, compute the relevant measures through network analysis and then determine which genres are the most infleuntials.